python/requirements-dev.txt
# Development dependencies for agentize Python modules
pytest>=7.0.0
pyyaml>=6.0
